Product page functionality/accessible from home
 Added functionality to product pages.
For example: /products/1/ links to the
product page for that corresponding book in the database. The homepage links to the first book (product) in the database.
The product page displays the title of the book, a brief description, the price, tags, and an any images uploaded to the images json field. When creating a product through the admin database, the images json field should look something like ["---url to any image---"].

I also added some default parameters to the stripe keys so the site can be accessed without issues. However, if you are wanting to test stripe, go in the root settings file and remove the default parameters